requirements.txt
Flask
dnspython
python-whois
requests
ipwhois
pyOpenSSL
